tech: webpack + karma, all tests pass
