+++ karma.conf.js
@@ -48,7 +48,7 @@ module.exports = function(config) {
         'react/addons': true,
         'react/lib/ExecutionEnvironment': true,
         'react/lib/ReactContext': true,
-        'systemjs/dist/system.src': true,
+        '.': true,
       },
       node: {
         fs: 'empty'
