+++ public/test/index.ts
@@ -7,8 +7,6 @@ import 'jquery';
 import angular from 'angular';
 import 'angular-mocks';
 import 'app/app';
-import System from 'systemjs/dist/system.src';
-console.log(System);
 // import './specs/test_specs';
 //
 // declare var window: any;
@@ -26,7 +24,7 @@ angular.module('grafana.directives', []);
 angular.module('grafana.filters', []);
 angular.module('grafana.routes', ['ngRoute']);
 
-const context = require.context('../', true, /core.*_specs\.ts/);
+const context = require.context('../', true, /.*_specs\.(tsx?|js)$/);
 
 for (let key of context.keys()) {
   console.log('key: ', key);
